my code
In this code, we declare two variables: number to store the user's input and factorial to store the factorial of the number.

We prompt the user to enter a positive integer using std::cout and read the input into the number variable using std::cin.

Next, we calculate the factorial of the number using a for loop. Starting from 1, we multiply factorial by each number from 1 up to and including the number entered by the user.